+++ apidaora/responses.py
@@ -1,5 +1,5 @@
 from http import HTTPStatus
-from typing import Any, Optional, Sequence
+from typing import Any, Dict, Optional, Sequence
 
 from dictdaora import DictDaora
 
@@ -11,18 +11,21 @@ class Response(DictDaora):
     status: HTTPStatus
     headers: Sequence[Header]
     content_type: Optional[ContentType]
+    kwargs: Optional[Dict[str, Any]]
 
 
 def json(
     body: Any,
     status: HTTPStatus = HTTPStatus.OK,
     headers: Sequence[Header] = (),
+    **kwargs: Any,
 ) -> Response:
     return Response(
         body=body,
         status=status,
         headers=headers,
         content_type=ContentType.APPLICATION_JSON,
+        kwargs=kwargs,
     )
 
 
@@ -30,12 +33,14 @@ def text(
     body: Any,
     status: HTTPStatus = HTTPStatus.OK,
     headers: Sequence[Header] = (),
+    **kwargs: Any,
 ) -> Response:
     return Response(
         body=body,
         status=status,
         headers=headers,
         content_type=ContentType.TEXT_PLAIN,
+        kwargs=kwargs,
     )
 
 
@@ -43,21 +48,24 @@ def html(
     body: Any,
     status: HTTPStatus = HTTPStatus.OK,
     headers: Sequence[Header] = (),
+    **kwargs: Any,
 ) -> Response:
     return Response(
         body=body,
         status=status,
         headers=headers,
         content_type=ContentType.TEXT_HTML,
+        kwargs=kwargs,
     )
 
 
-def no_content(headers: Sequence[Header] = ()) -> Response:
+def no_content(headers: Sequence[Header] = (), **kwargs: Any) -> Response:
     return Response(
         status=HTTPStatus.NO_CONTENT,
         headers=headers,
         content_type=None,
         body=None,
+        kwargs=kwargs,
     )
 
 
@@ -65,10 +73,12 @@ def not_found(
     body: Optional[Any] = None,
     headers: Sequence[Header] = (),
     content_type: Optional[ContentType] = None,
+    **kwargs: Any,
 ) -> Response:
     return Response(
         body=body,
         status=HTTPStatus.NOT_FOUND,
         headers=headers,
         content_type=content_type,
+        kwargs=kwargs,
     )